Separate workflow service tests from document processing tests
• Extracted workflow tests into dedicated WorkflowServiceTests class
• Simplified test setup by removing DocumentProcessingService dependency
• Improved test isolation by manually initializing document pages
• Added more assertions to cover additional test cases
• Reduced test complexity by removing actual file processing dependencies
